bess-marketdata-ingestion/providers/mengxi/requirements.txt
requests>=2.28
python-dotenv>=1.0
tqdm>=4.65
pandas
sqlalchemy
psycopg2-binary
openpyxl
numpy
numpy-financial